requirements.txt
moviepy==1.0.3
Pillow==10.0.0
numpy==1.24.3
scipy==1.11.1
python-dotenv==1.0.0
requests==2.31.0
